requirements.txt
python-dotenv
google-adk
litellm
google-adk[eval]